add calculation to determine step count by checking difference of last step readData() call and subtracting that from most recent readData() call.

+++ StrideRPG/app/src/main/java/stride/com/striderpg/fit/FitnessUtil.java
@@ -51,7 +51,7 @@ public class FitnessUtil {
                         new OnSuccessListener<DataSet>() {
                             @Override
                             public void onSuccess(DataSet dataSet) {
-                                long total =
+                                Integer total =
                                         dataSet.isEmpty()
                                                 ? 0
                                                 : dataSet.getDataPoints().get(0).getValue(Field.FIELD_STEPS).asInt();
